add detail/conversions/to_json.hpp
